+++ src/app/shared/fuzzy-search.function.ts
@@ -1,0 +1,21 @@
+// https://github.com/bevacqua/fuzzysearch
+export function fuzzySearch(needle, haystack) {
+  let hlen = haystack.length;
+  let nlen = needle.length;
+  if (nlen > hlen) {
+    return false;
+  }
+  if (nlen === hlen) {
+    return needle === haystack;
+  }
+  outer: for (let i = 0, j = 0; i < nlen; i++) {
+    let nch = needle.charCodeAt(i);
+    while (j < hlen) {
+      if (haystack.charCodeAt(j++) === nch) {
+        continue outer;
+      }
+    }
+    return false;
+  }
+  return true;
+}